requirements.txt
pyyaml>=6.0
pandas>=2.0.0
numpy>=1.24.0
duckdb>=0.9.0
redis>=5.0.0
asyncua>=1.0.0
PyQt6>=6.0.0
matplotlib>=3.7.0

